1. Run `test_preprocess.py` in ci.
2. Refactor `test_preprocess.py` by creating the model artifacts (eager, exported, et) once and re-use them on different inputs (previously, create model artifacts for each test).

Test Plan:
```
pytest -c /dev/null -v -n auto examples/models/llama3_2_vision/preprocess/
```

[ghstack-poisoned]

+++ backends/cadence/hifi/operators/dequantize_per_tensor.cpp
@@ -45,7 +45,10 @@ void dequantize_per_tensor_out(
     const int32_t* input_data = input.const_data_ptr<int32_t>();
     dequantize<int32_t>(out_data, input_data, scale, zero_point, numel);
   } else {
-    ET_CHECK_MSG(false, "Unhandled input dtype %hhd", input.scalar_type());
+    ET_CHECK_MSG(
+        false,
+        "Unhandled input dtype %hhd",
+        static_cast<int8_t>(input.scalar_type()));
   }
 }
